refactor: move yellow side element on layout
